finish wireless and mqtt

+++ application/pointManager.h
@@ -2,9 +2,10 @@
 #define __POINTMANAGER_H_
 
 
-void setPointId(int id);
+void setPointId(short id);
 void loopPointHeart(void);
 int getPointIDData(unsigned char *buff);
+
 #endif
